+++ ilivery/layers/__init__.py
@@ -8,6 +8,8 @@ from .texture_layer import texture_layer
 from .class_decal_layer import class_decal_layer
 from .patch_layer import patch_layer
 from .pattern_layer import pattern_layer
+from .stripe_layer import stripe_layer
+from .psd_layer import psd_layer
 
 from ilivery.layer import Layer
 
@@ -18,6 +20,8 @@ _layer_dict = {
     "CLASS_DECAL": class_decal_layer,
     "PATCH": patch_layer,
     "PATTERN": pattern_layer,
+    "STRIPE": stripe_layer,
+    "PSD": psd_layer,
 }