+++ tests/layers/test_solid_layer.py
@@ -1,10 +1,7 @@
 import pydantic
 
-from ilivery.layer import Layer
 from ilivery.layers import layer_from_config
 from ilivery.config import layer_configs
-
-import pytest
 
 
 class TestSolidLayer: