+++ tests/layers/test_texture_layer.py
@@ -1,11 +1,8 @@
 #!/usr/bin/env python3
 import pydantic
 
-from ilivery.layer import Layer
 from ilivery.layers import layer_from_config
 from ilivery.config import layer_configs
-
-import pytest
 
 
 class TestTextureLayer: