+++ tests/test_colorfuncs.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+import os
 import pydantic
 import numpy as np
 import matplotlib as mpl
@@ -44,6 +45,7 @@ class TestGradientCFunc:
 
         compare_ref_image(image)
 
+    @pytest.mark.skipif(os.getenv("GITHUB_ACTIONS") == "true", reason="Fails in GHAction env")
     def test_basic_xy(self, compare_ref_image):
         config = {
             "type": "GRADIENT",
@@ -54,6 +56,38 @@ class TestGradientCFunc:
         cfunc = colorfunc_from_config(config)
 
         image = _image_from_cfunc(cfunc)
+
+        compare_ref_image(image)
+
+
+class TestVoronoiCFunc:
+    def test_basic(self, compare_ref_image):
+        config = {
+            "type": "VORONOI",
+            "levels": 3,
+            "spacing": 20,
+        }
+
+        config = pydantic.TypeAdapter(ColorFunction).validate_python(config)
+        cfunc = colorfunc_from_config(config)
+
+        image = _image_from_cfunc(cfunc)
+
+        compare_ref_image(image)
+
+    def test_basic_anisotropic(self, compare_ref_image):
+        config = {
+            "type": "VORONOI",
+            "levels": 3,
+            "spacing": 20,
+            "anisotropy": 4,
+            "angle": 15,
+        }
+
+        config = pydantic.TypeAdapter(ColorFunction).validate_python(config)
+        cfunc = colorfunc_from_config(config)
+
+        image = _image_from_cfunc(cfunc, size=(200, 200))
 
         compare_ref_image(image)