+++ docs/CHANGELOG.md
@@ -1,3 +1,17 @@
+# v0.2.0 [UNRELEASED]
+
+- Added
+  - #2 Add hex layer
+  - #2 Add Voronoi noise
+  - #2 Add stripe layer
+  - #2 Add PSD layer
+ 
+- Fixed
+  - #2 Fix issue with PSDs with small layers
+ 
+- Changed
+  - #2 Switch to UV
+
 # v0.1.0 [2024-08-17]
 
 Initial release